requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
python-decouple==3.8
sqlparse==0.5.3
djangorestframework-simplejwt==5.5.0
pyTelegramBotAPI==4.27.0
psycopg2-binary==2.9.10
redis==6.2.0
celery==5.5.3
drf-spectacular==0.28.0
django-filter==25.1
geopy==2.4.1
django-modeltranslation==0.19.15
daphne==4.2.0
channels==4.2.2
channels_redis==4.2.1
django-leaflet==0.32.0
Pillow==11.3.0
django-query-counter==0.4.1
matplotlib==3.10.3
pandas==2.3.1
Faker==37.4.0
requests~=2.32.4
djangorestframework-gis~=1.2.0
PyJWT~=2.9.0
httpx==0.28.1
django-redis==6.0.0
pytest==8.4.1
pytest-django==4.11.1